Split BitmapFont and FontManager into two  modules
