+++ doc/modules/ext/bitmapfont.rst
@@ -1,0 +1,8 @@
+`sdl2.ext.bitmapfont` - Basic Bitmap Font Rendering
+===================================================
+
+This module provides the :class:`~sdl2.ext.BitmapFont` class, which allows for
+basic font rendering in PySDL2 without depending on the SDL_ttf library.
+
+.. automodule:: sdl2.ext.bitmapfont
+	:members:
