--- doc/modules/ext/font.rst
@@ -1,7 +1,0 @@
-.. currentmodule:: sdl2.ext
-
-Text rendering routines
-=======================
-
-.. automodule:: sdl2.ext.font
-	:members:
