+++ doc/modules/ext/ttf.rst
@@ -1,0 +1,7 @@
+.. currentmodule:: sdl2.ext
+
+TrueType Text rendering routines
+================================
+
+.. automodule:: sdl2.ext.ttf
+	:members:
