+++ sdl2/ext/__init__.py
@@ -13,8 +13,9 @@ from .ebs import *
 
 from .common import *
 from .draw import *
-from .font import *
-from .gui import *
+from .bitmapfont import *
+from .ttf import *
+from .uisystem import *
 from .msgbox import *
 from .image import *
 from .pixelaccess import *
